DEV-3151:  dcap-rs refactoring

+++ src/types/cert.rs
@@ -3,28 +3,28 @@ use x509_parser::{certificate::X509Certificate, revocation_list::CertificateRevo
 
 use crate::utils::cert::{get_crl_uri, is_cert_revoked, parse_x509_der_multi, pem_to_der};
 
-use super::IntelCollateral;
+use super::collaterals::IntelCollateral;
 
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct SgxExtensionTcbLevel {
-    pub sgxtcbcomp01svn: u64,
-    pub sgxtcbcomp02svn: u64,
-    pub sgxtcbcomp03svn: u64,
-    pub sgxtcbcomp04svn: u64,
-    pub sgxtcbcomp05svn: u64,
-    pub sgxtcbcomp06svn: u64,
-    pub sgxtcbcomp07svn: u64,
-    pub sgxtcbcomp08svn: u64,
-    pub sgxtcbcomp09svn: u64,
-    pub sgxtcbcomp10svn: u64,
-    pub sgxtcbcomp11svn: u64,
-    pub sgxtcbcomp12svn: u64,
-    pub sgxtcbcomp13svn: u64,
-    pub sgxtcbcomp14svn: u64,
-    pub sgxtcbcomp15svn: u64,
-    pub sgxtcbcomp16svn: u64,
-    pub pcesvn: u64,
+    pub sgxtcbcomp01svn: u8,
+    pub sgxtcbcomp02svn: u8,
+    pub sgxtcbcomp03svn: u8,
+    pub sgxtcbcomp04svn: u8,
+    pub sgxtcbcomp05svn: u8,
+    pub sgxtcbcomp06svn: u8,
+    pub sgxtcbcomp07svn: u8,
+    pub sgxtcbcomp08svn: u8,
+    pub sgxtcbcomp09svn: u8,
+    pub sgxtcbcomp10svn: u8,
+    pub sgxtcbcomp11svn: u8,
+    pub sgxtcbcomp12svn: u8,
+    pub sgxtcbcomp13svn: u8,
+    pub sgxtcbcomp14svn: u8,
+    pub sgxtcbcomp15svn: u8,
+    pub sgxtcbcomp16svn: u8,
+    pub pcesvn: u16,
     pub cpusvn: [u8; 16]
 }
